backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.125.0",
    "google-generativeai>=0.8.6",
    "litellm>=1.80.12",
    "opencv-python>=4.12.0.88",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.21",
    "ruff>=0.14.10",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pymupdf>=1.26.7",
    "greenlet>=3.3.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pyre-check>=0.9.25",
    "pyrefly>=0.49.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "types-requests>=2.32.4.20260107",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"] # Enable isort (I)
ignore = []

[tool.pyrefly]
project-includes = ["src"]
ignore-missing-imports = ["*"]
